Iter5 end 2025 05 08

+++ app/controllers/courses_controller.rb
@@ -1,10 +1,11 @@
 class CoursesController < ApplicationController
   before_action :authenticate_user
-  before_action :set_course, only: %i[show edit sync_assignments sync_enrollments enrollments]
+  before_action :set_course, only: %i[show edit sync_assignments sync_enrollments enrollments delete]
   before_action :set_pending_request_count
   before_action :determine_user_role
 
   def index
+    # To-Do: lms creation shouldn't be in this controller, it should probably be in the earliest set up stages
     Lms.find_or_create_by(id: 1, lms_name: 'Canvas', use_auth_token: true)
     @teacher_courses = UserToCourse.includes(:course).where(user: @user, role: %w[teacher ta])
 
@@ -78,41 +79,24 @@ class CoursesController < ApplicationController
     @side_nav = 'enrollments'
     return redirect_to courses_path, alert: 'You do not have access to this page.' unless @role == 'instructor'
 
-    @enrollments = @course.user_to_courses.includes(:user).sort_by do |enrollment|
-      role_priority = case enrollment.role
-                      when 'teacher' then 0
-                      when 'ta' then 1
-                      when 'student' then 2
-                      else 3
-                      end
-      [role_priority, enrollment.user.name.downcase]
-    end
+    @enrollments = @course.user_to_courses.includes(:user)
   end
 
-  # DELETES ALL COURSES, ASSIGNMENTS, EXTENSIONS, AND USER-TO-COURSE MAPPINGS
-  # This method is intended for use in development or testing environments only.
-  def delete_all
-    user_courses = Course.joins(:user_to_courses).where(user_to_courses: { user_id: @user.id })
+  def delete
+    return redirect_to courses_path, alert: 'You do not have access to this page.' unless @role == 'instructor'
+    return redirect_to courses_path, alert: 'Extensions are enabled for this course.' if @course.course_settings&.enable_extensions
 
-    # Find all assignments associated with the user's courses
-    assignments = Assignment.where(course_to_lms_id: CourseToLms.where(course_id: user_courses).select(:id))
-
-    # Delete all extensions associated with those assignments
+    assignments = Assignment.where(course_to_lms_id: CourseToLms.where(course_id: @course.id).select(:id))
     Extension.where(assignment_id: assignments.select(:id)).destroy_all
-
-    # Delete the assignments, course-to-LMS mappings, and user-to-course mappings
     assignments.destroy_all
-    CourseToLms.where(course_id: user_courses).destroy_all
-    UserToCourse.where(course_id: user_courses).destroy_all
-
-    # Delete course_settings and form_settings associated with the user's courses
-    CourseSettings.where(course_id: user_courses).destroy_all
-    FormSetting.where(course_id: user_courses).destroy_all
-
-    # Delete courses that no longer have any user-to-course associations
+    CourseToLms.where(course_id: @course.id).destroy_all
+    UserToCourse.where(course_id: @course.id).destroy_all
+    Request.where(course_id: @course.id).destroy_all
+    CourseSettings.where(course_id: @course.id).destroy_all
+    FormSetting.where(course_id: @course.id).destroy_all
     Course.where.missing(:user_to_courses).destroy_all
 
-    redirect_to courses_path, notice: 'All your courses and associations have been deleted successfully.'
+    redirect_to courses_path, notice: 'Course deleted successfully.'
   end
 
   private
@@ -129,14 +113,6 @@ class CoursesController < ApplicationController
 
   def determine_user_role
     @role = @course&.user_role(@user)
-  end
-
-  def render_role_based_view(instructor_view = 'courses/instructor_view', student_view = 'courses/student_view')
-    case @role
-    when 'instructor' then render instructor_view
-    when 'student' then render student_view
-    else redirect_to courses_path, alert: 'You do not have access to this course.'
-    end
   end
 
   def filter_courses(courses, roles, exclude_ids = [])
